Refactor package names and Add owner to Deed
