+++ app/src/main/java/com/ss20/se2/monopoly/models/fields/cards/Card.java
@@ -1,4 +1,6 @@
-package com.ss20.se2.monopoly.models.fields;
+package com.ss20.se2.monopoly.models.fields.cards;
+
+import com.ss20.se2.monopoly.models.fields.GameTile;
 
 public abstract class Card extends GameTile{
