+++ app/src/main/java/com/ss20/se2/monopoly/models/fields/cards/ChanceCard.java
@@ -1,4 +1,4 @@
-package com.ss20.se2.monopoly.models.fields;
+package com.ss20.se2.monopoly.models.fields.cards;
 
 public class ChanceCard extends Card{
