+++ app/src/main/java/com/ss20/se2/monopoly/models/fields/deeds/Deed.java
@@ -1,4 +1,8 @@
-package com.ss20.se2.monopoly.models.fields;
+package com.ss20.se2.monopoly.models.fields.deeds;
+
+import com.ss20.se2.monopoly.models.Player;
+import com.ss20.se2.monopoly.models.fields.GameTile;
+
 /*
  * Deed definition:
  * A deed for each property is given to a player to signify ownership, and specifies purchase price, mortgage value,
@@ -11,6 +15,7 @@ public abstract class Deed extends GameTile{
 
 	private int price;
 	private int mortgage;
+	private Player owner;
 
 	Deed(String name, int price, int mortgage){
 		super(name);
@@ -24,5 +29,13 @@ public abstract class Deed extends GameTile{
 
 	public int getMortgage(){
 		return mortgage;
+	}
+
+	public Player getOwner(){
+		return owner;
+	}
+
+	public void setOwner(Player owner){
+		this.owner = owner;
 	}
 }
